fixed fastapi and restructured the api
